backend/requirements.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis and Celery
redis==5.0.1
celery==5.3.6

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# DICOM processing
pydicom==2.4.4
pynetdicom==2.0.2

# Image processing
Pillow==10.2.0
numpy==1.26.3

# HTTP client
httpx==0.26.0
aiohttp==3.9.1

# LLM clients
openai==1.9.0
anthropic==0.14.0
google-generativeai==0.3.2

# Monitoring
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
structlog==24.1.0
